requirements.txt
Flask==2.1.3
torch==1.10.1
opencv-python-headless==4.5.5.62
numpy==1.21.5
Pillow==8.4.0
matplotlib==3.5.1
joblib==1.1.1
